pyproject.toml
[project]
name = "spike-prime-upload"
version = "0.1.0"
description = ""
authors = [
    {name = "JoJoJux",email = "80523427+J0J0HA@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13.1,<4.0"
dependencies = [
    "compyner (>=0.2.0,<0.3.0)",
    "pyserial (>=3.5,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "json-fix (>=1.0.0,<2.0.0)",
    "dataclasses-json (>=0.6.7,<1.0.0)",
]

[project.urls]
repository = "https://github.com/GSG-Robots/spike-prime-upload"

[project.scripts]
spike-prime-upload = "spike_prime_upload.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
